+++ cap-async-std/src/fs/dir_entry.rs
@@ -1,9 +1,6 @@
-use crate::{
-    fs::{FileType, Metadata},
-    sys,
-};
+use crate::fs::{Dir, FileType, Metadata};
 use async_std::io;
-use std::{ffi, fmt};
+use std::{ffi, fmt, path::PathBuf};
 
 /// Entries returned by the `ReadDir` iterator.
 ///
@@ -20,7 +17,17 @@ use std::{ffi, fmt};
 ///
 /// [`std::fs::DirEntry`]: https://doc.rust-lang.org/std/fs/struct.DirEntry.html
 pub struct DirEntry<'dir> {
-    sys: sys::fs::DirEntry<'dir>,
+    dir: &'dir Dir,
+    name: PathBuf,
+    file_type: FileType,
+    #[cfg(any(
+        unix,
+        target_os = "redox",
+        target_os = "vxworks",
+        target_os = "fuchsia",
+        target_os = "wasi"
+    ))]
+    ino: u64,
 }
 
 impl<'dir> DirEntry<'dir> {
@@ -31,7 +38,7 @@ impl<'dir> DirEntry<'dir> {
     /// [`std::fs::DirEntry::metadata`]: https://doc.rust-lang.org/std/fs/struct.DirEntry.html#method.metadata
     #[inline]
     pub fn metadata(&self) -> io::Result<Metadata> {
-        self.sys.metadata()
+        self.dir.metadata(&self.name)
     }
 
     /// Returns the file type for the file that this entry points at.
@@ -41,7 +48,7 @@ impl<'dir> DirEntry<'dir> {
     /// [`std::fs::DirEntry::file_type`]: https://doc.rust-lang.org/std/fs/struct.DirEntry.html#method.file_type
     #[inline]
     pub fn file_type(&self) -> io::Result<FileType> {
-        self.sys.file_type()
+        Ok(self.file_type)
     }
 
     /// Returns the bare file name of this directory entry without any other leading path component.
@@ -51,7 +58,7 @@ impl<'dir> DirEntry<'dir> {
     /// [`std::fs::DirEntry::file_name`]: https://doc.rust-lang.org/std/fs/struct.DirEntry.html#method.file_name
     #[inline]
     pub fn file_name(&self) -> ffi::OsString {
-        self.sys.file_name()
+        self.name.clone().into_os_string()
     }
 }
 
@@ -59,7 +66,7 @@ impl<'dir> DirEntry<'dir> {
 impl<'dir> async_std::os::unix::fs::DirEntryExt for DirEntry<'dir> {
     #[inline]
     fn ino(&self) -> u64 {
-        self.sys.ino()
+        self.ino
     }
 }
