+++ cap-async-std/src/lib.rs
@@ -28,8 +28,6 @@
 #![deny(missing_docs)]
 #![cfg_attr(target_os = "wasi", feature(wasi_ext))]
 
-mod sys;
-
 pub mod fs;
 #[cfg(feature = "fs_utf8")]
 pub mod fs_utf8;
