--- cap-async-std/src/sys/wasi/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod fs;
