--- cap-async-std/src/sys/yanix/fs/mod.rs
@@ -1,7 +1,0 @@
-mod dir;
-mod dir_entry;
-mod read_dir;
-
-pub(crate) use dir::*;
-pub(crate) use dir_entry::*;
-pub(crate) use read_dir::*;
